refactor(github): Reorganize adapters
