+++ lib/util/github/graphql/datasource-fetcher.spec.ts
@@ -1,5 +1,4 @@
 import AggregateError from 'aggregate-error';
-import { parse as graphqlParse } from 'graphql';
 import * as httpMock from '../../../../test/http-mock';
 import { GithubGraphqlResponse, GithubHttp } from '../../http/github';
 import { range } from '../../range';
@@ -92,24 +91,6 @@ async function catchError<T>(cb: () => Promise<T>): Promise<Error> {
 }
 
 describe('util/github/graphql/datasource-fetcher', () => {
-  describe('prepareQuery', () => {
-    it('returns valid query for valid payload query', () => {
-      const payloadQuery = adapter.query;
-      expect(() => graphqlParse(`query { ${payloadQuery} }`)).not.toThrow();
-      expect(() =>
-        graphqlParse(Datasource.prepareQuery(payloadQuery))
-      ).not.toThrow();
-    });
-
-    it('returns invalid query for invalid payload query', () => {
-      const payloadQuery = '!@#';
-      expect(() => graphqlParse(`query { ${payloadQuery} }`)).toThrow();
-      expect(() =>
-        graphqlParse(Datasource.prepareQuery(payloadQuery))
-      ).toThrow();
-    });
-  });
-
   describe('query', () => {
     let http = new GithubHttp();
